+++ apps/zhi-cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # zhi-cli
 
+## 1.5.15
+
+### Patch Changes
+
+- zhi-log@1.14.15
+
+## 1.5.14
+
+### Patch Changes
+
+- zhi-log@1.14.14
+
+## 1.5.13
+
+### Patch Changes
+
+- zhi-log@1.14.13
+
 ## 1.5.12
 
 ### Patch Changes