+++ libs/zhi-cmd/CHANGELOG.md
@@ -1,5 +1,38 @@
 # zhi-cmd
 
+## 0.3.0
+
+### Minor Changes
+
+- feat: add MainFunction
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - zhi-lib-base@0.6.0
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.6.0
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.5.0
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.4.0
+
 ## 0.2.0
 
 ### Minor Changes