+++ libs/zhi-infra/CHANGELOG.md
@@ -1,5 +1,52 @@
 # zhi-infra
 
+## 0.5.0
+
+### Minor Changes
+
+- feat: add MainFunction
+- feat: add MainFunction
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - zhi-lib-base@0.6.0
+  - zhi-common@1.24.1
+
+## 0.4.0
+
+### Minor Changes
+
+- feat: add params
+
+## 0.3.0
+
+### Minor Changes
+
+- feat: change zhi infra params to array
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.6.0
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.5.0
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.4.0
+
 ## 0.2.0
 
 ### Minor Changes