+++ libs/zhi-log/CHANGELOG.md
@@ -1,5 +1,26 @@
 # zhi-log
 
+## 1.14.15
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.6.0
+
+## 1.14.14
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.5.0
+
+## 1.14.13
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-device@2.4.0
+
 ## 1.14.12
 
 ### Patch Changes